Updated sprites, began making a new intro, fixed bugs, improved text
Improved the Robo sprites.
Got rid of empty gamestartup.ccs file
Started making the new intro.
Got rid of "This is a room for guests" popup, as it doesn't make much
sense.
Actually made some text appear.
Improved the writing of thoughts.
Made the sprites not show during the flyover.